Рефакторинг, разделит форму комментария на более мелкие компоненты
