+++ project/src/const.js
@@ -65,3 +65,26 @@ export const SortType = {
 };
 
 export const DEFAULT_SORT_TYPE = SortType.POPULAR;
+
+export const Ratings = {
+  PERFECT: {
+    value: 5,
+    title: 'perfect',
+  },
+  GOOD: {
+    value: 4,
+    title: 'good',
+  },
+  NOT_BAD: {
+    value: 3,
+    title: 'not bad',
+  },
+  BADLY: {
+    value: 2,
+    title: 'badly',
+  },
+  TERRIBLY: {
+    value: 1,
+    title: 'terribly',
+  },
+};
